% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{as.data.frame.moderncor}
\alias{as.data.frame.moderncor}
\title{Convert moderncor object to a data.frame}
\usage{
\method{as.data.frame}{moderncor}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{An object of class "moderncor".}

\item{row.names}{Ignored.}

\item{optional}{Ignored.}

\item{...}{Additional arguments.}
}
\value{
A data.frame.
}
\description{
Convert moderncor object to a data.frame
}
