% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_cat.R
\name{as.data.frame.moderncor_cat}
\alias{as.data.frame.moderncor_cat}
\title{Convert moderncor_cat object to a data.frame}
\usage{
\method{as.data.frame}{moderncor_cat}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{An object of class "moderncor_cat".}

\item{row.names}{Ignored.}

\item{optional}{Ignored.}

\item{...}{Additional arguments.}
}
\value{
A data.frame.
}
\description{
Convert moderncor_cat object to a data.frame
}
