% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moderncor_cat.R
\name{available_methods_cat}
\alias{available_methods_cat}
\title{List all available categorical association methods}
\usage{
available_methods_cat()
}
\value{
A data.frame with method names, labels, package requirements, and type.
}
\description{
List all available categorical association methods
}
\examples{
available_methods_cat()
}
