% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_suggested}
\alias{check_suggested}
\title{Check if a suggested package is available}
\usage{
check_suggested(pkg, method)
}
\arguments{
\item{pkg}{The package name.}

\item{method}{The method requesting the package.}
}
\description{
Check if a suggested package is available
}
\keyword{internal}
