% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_matrix}
\alias{compute_matrix}
\title{Compute correlation matrix for all variable pairs}
\usage{
compute_matrix(
  mat,
  z = NULL,
  method,
  use,
  alternative,
  p_value,
  call,
  method_partial = "pearson",
  ...
)
}
\arguments{
\item{mat}{Numeric matrix.}

\item{z}{Numeric vector, matrix, or data.frame representing control variables.}

\item{method}{Character: correlation method.}

\item{use}{Character: missing value treatment.}

\item{alternative}{Character: alternative hypothesis.}

\item{p_value}{Logical: whether to compute p-values.}

\item{call}{The original function call.}

\item{method_partial}{Character: correlation method to use for partial/semi-partial.}

\item{...}{Additional arguments.}
}
\value{
A moderncor object for matrix input.
}
\description{
Compute correlation matrix for all variable pairs
}
\keyword{internal}
