% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moderncor_cat.R
\name{compute_matrix_cat}
\alias{compute_matrix_cat}
\title{Compute categorical correlation matrix for all variable pairs}
\usage{
compute_matrix_cat(df, method, use, call, ...)
}
\description{
Compute categorical correlation matrix for all variable pairs
}
\keyword{internal}
