% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_information.R
\name{compute_mutual_info}
\alias{compute_mutual_info}
\title{Compute Mutual Information}
\usage{
compute_mutual_info(x, y, alternative = "two.sided", p_value = TRUE, ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{y}{Numeric vector.}

\item{alternative}{Character: alternative hypothesis (ignored with warning if not "two.sided").}

\item{p_value}{Logical: whether to compute p-value via permutation.}

\item{...}{Additional arguments passed to infotheo::mutinformation or discretization.}
}
\value{
A list of correlation results.
}
\description{
Compute Mutual Information
}
\keyword{internal}
