% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_pair}
\alias{compute_pair}
\title{Dispatch to the appropriate pair-wise compute function}
\usage{
compute_pair(x, y, z = NULL, method, alternative, p_value, ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{y}{Numeric vector.}

\item{z}{Numeric vector, matrix, or data.frame representing control variables.}

\item{method}{Character: correlation method.}

\item{alternative}{Character: alternative hypothesis.}

\item{p_value}{Logical: whether to calculate p-value.}

\item{...}{Additional arguments.}
}
\value{
A list of correlation results.
}
\description{
Dispatch to the appropriate pair-wise compute function
}
\keyword{internal}
