% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moderncor_cat.R
\name{compute_pair_cat}
\alias{compute_pair_cat}
\title{Dispatch to the appropriate pair-wise compute function for categorical variables}
\usage{
compute_pair_cat(x, y, method, ...)
}
\description{
Dispatch to the appropriate pair-wise compute function for categorical variables
}
\keyword{internal}
