% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_classic.R
\name{compute_pearson}
\alias{compute_pearson}
\title{Compute Pearson correlation coefficient}
\usage{
compute_pearson(x, y, alternative = "two.sided", p_value = TRUE, ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{y}{Numeric vector.}

\item{alternative}{Character: alternative hypothesis.}

\item{p_value}{Logical: whether to compute p-value.}

\item{...}{Additional arguments.}
}
\value{
A list of correlation results.
}
\description{
Compute Pearson correlation coefficient
}
\keyword{internal}
