% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_other.R
\name{compute_tau_star}
\alias{compute_tau_star}
\title{Compute Bergsma-Dassios Tau* Correlation}
\usage{
compute_tau_star(x, y, alternative = "two.sided", p_value = TRUE, ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{y}{Numeric vector.}

\item{alternative}{Character: alternative hypothesis (ignored with warning if not "two.sided").}

\item{p_value}{Logical: whether to compute p-value.}

\item{...}{Additional arguments passed to TauStar::tauStarTest.}
}
\value{
A list of correlation results.
}
\description{
Compute Bergsma-Dassios Tau* Correlation
}
\keyword{internal}
