% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_methods.R
\name{method_info}
\alias{method_info}
\title{Get detailed information about a specific correlation method}
\usage{
method_info(method)
}
\arguments{
\item{method}{Character: the method name.}
}
\value{
A list with method details.
}
\description{
Get detailed information about a specific correlation method
}
\examples{
method_info("pearson")
}
