% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moderncor_cat.R
\name{method_info_cat}
\alias{method_info_cat}
\title{Information about categorical association methods}
\usage{
method_info_cat(method)
}
\description{
Information about categorical association methods
}
\keyword{internal}
