% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.moderncor}
\alias{print.moderncor}
\title{Print a moderncor object}
\usage{
\method{print}{moderncor}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class "moderncor".}

\item{digits}{Integer: number of decimal places to print.}

\item{...}{Additional arguments.}
}
\value{
The input object invisibly.
}
\description{
Print a moderncor object
}
