% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary.moderncor}
\alias{print.summary.moderncor}
\title{Print a summary.moderncor object}
\usage{
\method{print}{summary.moderncor}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class "summary.moderncor".}

\item{digits}{Integer: number of decimal places to print.}

\item{...}{Additional arguments.}
}
\value{
The input object invisibly.
}
\description{
Print a summary.moderncor object
}
