% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_cat.R
\name{print.summary.moderncor_cat}
\alias{print.summary.moderncor_cat}
\title{Print a summary.moderncor_cat object}
\usage{
\method{print}{summary.moderncor_cat}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class "summary.moderncor_cat".}

\item{digits}{Integer: number of decimal places to print.}

\item{...}{Additional arguments.}
}
\value{
The input object invisibly.
}
\description{
Print a summary.moderncor_cat object
}
