% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.moderncor}
\alias{summary.moderncor}
\title{Summarize a moderncor object}
\usage{
\method{summary}{moderncor}(object, ...)
}
\arguments{
\item{object}{An object of class "moderncor".}

\item{...}{Additional arguments.}
}
\value{
An object of class "summary.moderncor".
}
\description{
Summarize a moderncor object
}
