% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_cat.R
\name{summary.moderncor_cat}
\alias{summary.moderncor_cat}
\title{Summarize a moderncor_cat object}
\usage{
\method{summary}{moderncor_cat}(object, ...)
}
\arguments{
\item{object}{An object of class "moderncor_cat".}

\item{...}{Additional arguments.}
}
\value{
An object of class "summary.moderncor_cat".
}
\description{
Summarize a moderncor_cat object
}
