% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moderncor_cat.R
\name{validate_input_cat}
\alias{validate_input_cat}
\title{Validate and normalize input for moderncor_cat}
\usage{
validate_input_cat(x, y, use)
}
\description{
Validate and normalize input for moderncor_cat
}
\keyword{internal}
