% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moderncor-package.R
\docType{package}
\name{moderncor-package}
\alias{moderncor-package}
\title{moderncor: Unified Interface for Modern and Classical Correlation Coefficients}
\description{
The moderncor package provides a single, unified interface to compute a wide range
of classical and modern correlation and association measures.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ToshihiroIguchi/moderncor}
  \item Report bugs at \url{https://github.com/ToshihiroIguchi/moderncor/issues}
}

}
\author{
\strong{Maintainer}: Toshihiro Iguchi \email{toshihiro.iguchi.github@gmail.com}

Authors:
\itemize{
  \item Toshihiro Iguchi \email{toshihiro.iguchi.github@gmail.com}
}

}
\keyword{internal}
